#104 - 장소 크롤링 기능 재활성화
